+++ Kits/clevertap/clevertap-7/Sources/mParticle-CleverTap/MPKitCleverTap.m
@@ -16,7 +16,7 @@ static NSString *const ctCleverTapIdIntegrationKey = @"clevertap_id_integration_
 static NSString *const kCTTransactionID = @"Transaction Id";
 static NSString *const kCTChargedID = @"Charged ID";
 static NSString *const kLibName = @"mParticle-iOS";
-static NSString *const kLibVersion = @"9.2.1";
+static NSString *const kLibVersion = @"9.2.2";
 
 @implementation MPKitCleverTap